Ruby Build
This change adds support for building Ruby runtimes to the dependency
builder.  It does this by using the ruby-build[1] project.

[1]: https://github.com/sstephenson/ruby-build

+++ lib/build/dependency/util/base_vagrant_platform.rb
@@ -25,9 +25,10 @@ module Build
         VAGRANT_PLATFORMS.include? platform
       end
 
-      def initialize(name, version)
-        @name    = name
-        @version = version
+      def initialize(name, version, shutdown)
+        @name     = name
+        @version  = version
+        @shutdown = shutdown
       end
 
       def exec(command)
@@ -37,7 +38,7 @@ module Build
           system "vagrant ssh #{@name} --command 'cd /java-buildpack-dependency-builder && #{command}'"
 
           abort 'FAILURE' unless $CHILD_STATUS == 0
-          system "vagrant destroy -f #{@name}"
+          system "vagrant destroy -f #{@name}" if @shutdown
         end
       end
